Fix helper test filename
Files that don't end with '-test', will not be run.
